refactor(planner): move the Moon-avoidance domain model out of features
Addresses the CodeRabbit review on #1596.

`shared/planner/guidance-settings.ts` imported from
`features/targets/astro/moon-avoidance`, while the target and settings
consumers import `shared/planner/guidance-settings`. That preserved a
feature -> shared -> feature edge, undermining the settings/targets decoupling
this branch exists to perform.

`moon-avoidance.ts` is pure domain logic with no imports of its own, and it is
read from both sides of that boundary, so `shared/planner/` is where it belongs.
Moved it (and its tests) there and pointed `guidance-settings.ts` at the local
path. The old `features/targets/astro/moon-avoidance` path stays as a re-export
shim so the ~14 call sites inside the targets feature are untouched.

Verified: `tsc --noEmit` clean, biome clean, the circular-dependency gate reports
the same three pre-existing allowlisted type-import cycles and no new ones, and
the full desktop suite passes (2284 tests).

Note for a follow-up, out of scope here: four other `shared -> features` imports
remain (`shared/settings-queries.ts`, `shared/observing-sites/ObservingSites.tsx`,
`shared/planner/catalogue-settings.ts`, `shared/planner/altitude-settings.ts`).
This commit removes only the edge the review identified.
