+++ apps/desktop/src/shared/planner/moon-avoidance.test.ts
